Cloud storage: correctly store pictures even if they weren't shown before
We could end up in a situation where the hash for a picture hadn't been
recorded yet and then the lookup to make sure that we find the correct
file actually got us nothing.

Now we make sure that the picture is in the hash table before looking up
the file location and before creating the name of the blob that we store
in git.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>