Teach copy_string to return NULL on empty string
In subsurface we store no value as a NULL pointer everywhere, but
sometimes some structures returns a empty field as a empty string.

This teaches our helper copy_string to return NULL if you try to copy a
empty string.

This fixes a bug where we store buddy and divemaster in git as empty
strings if they passed via MainTab::saveTaggedStrings().

Suggested-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Anton Lundin <glance@acc.umu.se>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>