Geo taxonomy: download the taxonomy data from geonames.org
There are a number of web servies we could use. All have their drawbacks.
This one is free with free data. It's daily limits are reasonably high.
For many coordinates I tested the results were good, for others at least
not terrible.

We can always consider supporting multiple such services. But for now this
seems like a reasonable choice.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>