Planner: correct output of planner mode
Previously we used strncat to output VPM mode without correctly defining the
length of the string, and didn't do anything for recreational mode.  This
resulted in the output being junk recycled from the previous temp string.

We could use strncat if the string length were defined, but using snprintf will
make it simpler to include the VPM conservatism when that has been implemented.

Signed-off-by: Rick Walsh <rickmwalsh@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>