Profile.c: Fix stepping through loop between sample points
If adjacent samples at times t0 and t1 are greater than time_stepsize apart,
this for loop steps through the time between samples to calculate tissue
tolerance incrementally at each intermediate time, j, (and interpolated depth).

If the difference between t1 and t0 is not a multiple of time_stepsize,
immediately before the final increment time_stepsize needs to be reduced to
t1 - j so that j = t1 in the final increment.

This is necessary when:
	t1 - j < time_stepsize	(i.e. we are about to start the final
				increment, and time_stepsize needs to be
				reduced), and
	j < t1			(i.e. we didn't just do the final increment -
				without this condition time_stepsize would be
				set to zero, leading to an infinite loop)

Previously, the check was (j - t0 < time_stepsize), which always returns false.

Signed-off-by: Rick Walsh <rickmwalsh@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>