Moved all core-functionality to subsurface-core
And adapted a new CMakeLists.txt file for it. on the way I'v also
found out that we where double-compilling a few files. I'v also
set the subsurface-core as a include_path but that was just to
reduce the noise on this commit, since I plan to remove it from
the include path to make it obligatory to specify something like

 include "subsurface-core/dive.h"

on the header files. Since the app is growing quite a bit we ended
up having a few different files with almost same name that did
similar things, I wanna kill that ( for instance Dive.h, dive.h,
PrintDive.h and such. )

Signed-off-by: Tomaz Canabrava <tomaz.canabrava@intel.com>