Hook the missing parts from the language prefs ui
And this actually made me find one Memory leak on the old version
We are constantly creating / leaking a new QSortFilterProxyModel
everytime a sync() is done on the settings. ( that's not a lot, maybe
once per time a user changes it's settings... but it's still a leak )

And the code is more sanely separated this time. yey.

Signed-off-by: Tomaz Canabrava <tomaz.canabrava@intel.com>